refactor: replace internal error references with new error package constants
